SERVER/requirements.txt
tensorflow
keras
tensorflow_hub
opencv-python
pillow